backend/requirements.txt
# Django core
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.0
djangorestframework-simplejwt>=5.3

# Database
psycopg2-binary>=2.9
dj-database-url>=2.0

# Production server
gunicorn>=21.0
whitenoise>=6.5

# Image handling
Pillow>=10.0

# HTTP client
requests>=2.31

# AI
google-generativeai>=0.3

# Utils
python-dotenv>=1.0
